requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser
drf-base64==2.0
django-extra-fields==3.0.2
Pillow==9.0.1
django-filter==22.1
